withmap: Access all user subuids
This is a tool to spawn a subprocess as the root user in a new user
namespace with uid/gid mappings according to commandline arguments.  The
actual mapping is done with the newuidmap and newgidmap utilities,
according to /etc/subuid and /etc/subgid, respectively.  This tool
itself is unprivileged.

This is useful for interacting with files from rootless podman
containers which may be inside your home directory.